requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
importlib-metadata==3.1.1
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
zipp==3.4.0
